backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.1
hedera-sdk-py==2.50.0
requests==2.31.0
email-validator==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0

